+++ src/VBox/Main/src-all/DisplayResampleImage.cpp
@@ -49,7 +49,7 @@
 #include <iprt/types.h>
 
 /* 2.0.10: cast instead of floor() yields 35% performance improvement.
-	Thanks to John Buckman. */
+        Thanks to John Buckman. */
 
 #define floor2(exp) ((long) exp)
 /*#define floor2(exp) floor(exp)*/
@@ -77,15 +77,15 @@ DECLINLINE(void) gdImageSetPixel (gdImagePtr im, int x, int y, int color, int w)
 #define gdTrueColorGetGreen(c) (((c) & 0x00FF00) >> 8)
 #define gdTrueColorGetBlue(c) ((c) & 0x0000FF)
 #define gdTrueColorAlpha(r, g, b, a) (((a) << 24) + \
-	((r) << 16) + \
-	((g) << 8) + \
-	(b))
+        ((r) << 16) + \
+        ((g) << 8) + \
+        (b))
 
 void gdImageCopyResampled (uint8_t *dst,
-		      uint8_t *src,
-		      int dstX, int dstY,
-		      int srcX, int srcY,
-		      int dstW, int dstH, int srcW, int srcH)
+                      uint8_t *src,
+                      int dstX, int dstY,
+                      int srcX, int srcY,
+                      int dstW, int dstH, int srcW, int srcH)
 {
   int x, y;
   double sy1, sy2, sx1, sx2;
@@ -93,105 +93,105 @@ void gdImageCopyResampled (uint8_t *dst,
     {
       sy1 = ((double) y - (double) dstY) * (double) srcH / (double) dstH;
       sy2 = ((double) (y + 1) - (double) dstY) * (double) srcH /
-	(double) dstH;
+        (double) dstH;
       for (x = dstX; (x < dstX + dstW); x++)
-	{
-	  double sx, sy;
-	  double spixels = 0;
-	  double red = 0.0, green = 0.0, blue = 0.0, alpha = 0.0;
-	  sx1 = ((double) x - (double) dstX) * (double) srcW / dstW;
-	  sx2 = ((double) (x + 1) - (double) dstX) * (double) srcW / dstW;
-	  sy = sy1;
-	  do
-	    {
-	      double yportion;
-	      if (floor2 (sy) == floor2 (sy1))
-		{
-		  yportion = 1.0 - (sy - floor2 (sy));
-		  if (yportion > sy2 - sy1)
-		    {
-		      yportion = sy2 - sy1;
-		    }
-		  sy = floor2 (sy);
-		}
-	      else if (sy == floor2 (sy2))
-		{
-		  yportion = sy2 - floor2 (sy2);
-		}
-	      else
-		{
-		  yportion = 1.0;
-		}
-	      sx = sx1;
-	      do
-		{
-		  double xportion;
-		  double pcontribution;
-		  int p;
-		  if (floor2 (sx) == floor2 (sx1))
-		    {
-		      xportion = 1.0 - (sx - floor2 (sx));
-		      if (xportion > sx2 - sx1)
-			{
-			  xportion = sx2 - sx1;
-			}
-		      sx = floor2 (sx);
-		    }
-		  else if (sx == floor2 (sx2))
-		    {
-		      xportion = sx2 - floor2 (sx2);
-		    }
-		  else
-		    {
-		      xportion = 1.0;
-		    }
-		  pcontribution = xportion * yportion;
-		  /* 2.08: previously srcX and srcY were ignored.
-		     Andrew Pattison */
-		  p = gdImageGetTrueColorPixel (src,
-						(int) sx + srcX,
-						(int) sy + srcY, srcW);
-		  red += gdTrueColorGetRed (p) * pcontribution;
-		  green += gdTrueColorGetGreen (p) * pcontribution;
-		  blue += gdTrueColorGetBlue (p) * pcontribution;
-		  alpha += gdTrueColorGetAlpha (p) * pcontribution;
-		  spixels += xportion * yportion;
-		  sx += 1.0;
-		}
-	      while (sx < sx2);
-	      sy += 1.0;
-	    }
-	  while (sy < sy2);
-	  if (spixels != 0.0)
-	    {
-	      red /= spixels;
-	      green /= spixels;
-	      blue /= spixels;
-	      alpha /= spixels;
-	    }
-	  /* Clamping to allow for rounding errors above */
-	  if (red > 255.0)
-	    {
-	      red = 255.0;
-	    }
-	  if (green > 255.0)
-	    {
-	      green = 255.0;
-	    }
-	  if (blue > 255.0)
-	    {
-	      blue = 255.0;
-	    }
-	  if (alpha > gdAlphaMax)
-	    {
-	      alpha = gdAlphaMax;
-	    }
-	  gdImageSetPixel (dst,
-			   x, y,
-			   gdTrueColorAlpha ((int) red,
-					     (int) green,
-					     (int) blue, (int) alpha), dstW);
-	}
+        {
+          double sx, sy;
+          double spixels = 0;
+          double red = 0.0, green = 0.0, blue = 0.0, alpha = 0.0;
+          sx1 = ((double) x - (double) dstX) * (double) srcW / dstW;
+          sx2 = ((double) (x + 1) - (double) dstX) * (double) srcW / dstW;
+          sy = sy1;
+          do
+            {
+              double yportion;
+              if (floor2 (sy) == floor2 (sy1))
+                {
+                  yportion = 1.0 - (sy - (double)floor2 (sy));
+                  if (yportion > sy2 - sy1)
+                    {
+                      yportion = sy2 - sy1;
+                    }
+                  sy = (double)floor2 (sy);
+                }
+              else if (sy == floor2 (sy2))
+                {
+                  yportion = sy2 - (double)floor2 (sy2);
+                }
+              else
+                {
+                  yportion = 1.0;
+                }
+              sx = sx1;
+              do
+                {
+                  double xportion;
+                  double pcontribution;
+                  int p;
+                  if (floor2 (sx) == floor2 (sx1))
+                    {
+                      xportion = 1.0 - (sx - (double)floor2 (sx));
+                      if (xportion > sx2 - sx1)
+                        {
+                          xportion = sx2 - sx1;
+                        }
+                      sx = (double)floor2 (sx);
+                    }
+                  else if (sx == floor2 (sx2))
+                    {
+                      xportion = sx2 - (double)floor2 (sx2);
+                    }
+                  else
+                    {
+                      xportion = 1.0;
+                    }
+                  pcontribution = xportion * yportion;
+                  /* 2.08: previously srcX and srcY were ignored.
+                     Andrew Pattison */
+                  p = gdImageGetTrueColorPixel (src,
+                                                (int) sx + srcX,
+                                                (int) sy + srcY, srcW);
+                  red += gdTrueColorGetRed (p) * pcontribution;
+                  green += gdTrueColorGetGreen (p) * pcontribution;
+                  blue += gdTrueColorGetBlue (p) * pcontribution;
+                  alpha += gdTrueColorGetAlpha (p) * pcontribution;
+                  spixels += xportion * yportion;
+                  sx += 1.0;
+                }
+              while (sx < sx2);
+              sy += 1.0;
+            }
+          while (sy < sy2);
+          if (spixels != 0.0)
+            {
+              red /= spixels;
+              green /= spixels;
+              blue /= spixels;
+              alpha /= spixels;
+            }
+          /* Clamping to allow for rounding errors above */
+          if (red > 255.0)
+            {
+              red = 255.0;
+            }
+          if (green > 255.0)
+            {
+              green = 255.0;
+            }
+          if (blue > 255.0)
+            {
+              blue = 255.0;
+            }
+          if (alpha > gdAlphaMax)
+            {
+              alpha = gdAlphaMax;
+            }
+          gdImageSetPixel (dst,
+                           x, y,
+                           gdTrueColorAlpha ((int) red,
+                                             (int) green,
+                                             (int) blue, (int) alpha), dstW);
+        }
     }
 }
